+++ lib/pages/intake_estimater.dart
@@ -278,7 +278,7 @@ class _IntakeEstimaterState extends State<IntakeEstimater> {
                     controller: selectedCalories,
                     min: provider.lowerCalorieRange.toDouble(),
                     max: provider.upperCalorieRange.toDouble(),
-                    sliderValue: provider.safeCurrentCalories,
+                    sliderValue: provider.safeCurrentCalories.toDouble(),
                     onSliderChanged: (newCurrent) =>
                         provider.setCurrentCalories(newCurrent),
                     onLongPressDecrement: null,
@@ -342,9 +342,9 @@ class _IntakeEstimaterState extends State<IntakeEstimater> {
                     title: '${provider.currentProtein.toStringAsFixed(0)} g',
                     textFieldHintText: 'Select Protein',
                     controller: proteinController,
-                    min: provider.lowerProteinRange,
-                    max: provider.upperProteinRange,
-                    sliderValue: provider.safeCurrentProtein,
+                    min: provider.lowerProteinRange.toDouble(),
+                    max: provider.upperProteinRange.toDouble(),
+                    sliderValue: provider.safeCurrentProtein.toDouble(),
                     onSliderChanged: (newCurrent) =>
                         provider.setCurrentProtein(newCurrent),
                     onLongPressDecrement: null,
